backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow>=11.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# AI Services
google-cloud-aiplatform==1.40.0
google-generativeai
numpy
requests
openai
fal-client==0.4.1

# Utilities
aiofiles==23.2.1
requests==2.31.0
